refactor: add generic gemini parser and nfd
