+++ examples/lammps-reax-rocky-8-mpich.json
@@ -1,0 +1,71 @@
+{
+    "software": [
+        "bash",
+        "dd",
+        "dnf:aspell",
+        "dnf:aspell-en",
+        "dnf:boost-devel",
+        "dnf:cmake",
+        "dnf:czmq-devel",
+        "dnf:development-tools",
+        "dnf:epel-release",
+        "dnf:fftw3",
+        "dnf:flux",
+        "dnf:git",
+        "dnf:grubby",
+        "dnf:hwloc",
+        "dnf:hwloc-devel",
+        "dnf:jansson-devel",
+        "dnf:jq",
+        "dnf:libarchive-devel",
+        "dnf:libedit-devel",
+        "dnf:libevent-devel",
+        "dnf:libxml2-devel",
+        "dnf:lua",
+        "dnf:lua-devel",
+        "dnf:lua-posix",
+        "dnf:lz4-devel",
+        "dnf:make",
+        "dnf:mpich",
+        "dnf:munge",
+        "dnf:munge-devel",
+        "dnf:ncurses-devel",
+        "dnf:nfs-utils",
+        "dnf:pmix",
+        "dnf:pmix-devel",
+        "dnf:python3-cffi",
+        "dnf:python3-devel",
+        "dnf:python3-docutils",
+        "dnf:python3-jsonschema",
+        "dnf:python3-sphinx",
+        "dnf:python3-yaml",
+        "dnf:sqlite-devel",
+        "dnf:valgrind-devel",
+        "dnf:wget",
+        "dnf:yaml-cpp-devel",
+        "flux-core",
+        "flux-pmix",
+        "flux-sched",
+        "flux-security",
+        "git",
+        "lammps",
+        "make",
+        "mpich",
+        "openpmix",
+        "perl",
+        "prrte",
+        "rh-scl:gcc-toolset-13",
+        "tar",
+        "yum:gcc-toolset-13"
+    ],
+    "devices": [
+        "urandom"
+    ],
+    "kernel": [
+        "cgroups-v2",
+        "systemd"
+    ],
+    "architecture": "amd64",
+    "uri": "ghcr.io/converged-computing/lammps-reax:rocky8-mpich"
+}
+
