+++ examples/lammps-reax-ubuntu-openmpi.json
